basics implemented for exerciseConteroller

+++ FATT_backend/WebAPI/Models/DishModel.cs
@@ -2,13 +2,12 @@
 {
     public class DishModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Category { get; set; }
         public double NutritunalValue { get; set; }
         public string Recipe { get; set; }
         public string Ingredients { get; set; }
         
-        // no clue what data type
         public string PicturePath { get; set; }
     }
 }
